feat: extract public servers page to separate component
